trying new approach that fixes code smells while not doing a deepClone of larger file uploads. Renamed common.js to uploadFileUtils.js for clarity
